test/com/ettp/OracleJDBCConnectionTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.ettp;

import java.sql.Connection;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author vdnh
 */
public class OracleJDBCConnectionTest {
    
    public OracleJDBCConnectionTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    /**
     * Test of getJDBCConnection method, of class OracleJDBCConnection.
     */
    @Test
    public void testGetJDBCConnection() {
        System.out.println("getJDBCConnection");
        //Connection expResult = null;
        Connection result = OracleJDBCConnection.getJDBCConnection();
        //assertEquals(expResult, result);
        assertNotNull(result);
        // TODO review the generated test code and remove the default call to fail.
        //fail("The test case is a prototype.");
    }
    
}
